fix(panel-admin): harden the engagement CSV and state the seat measurement
